Define a map's key reader for int64 type
The main library doesn't cater to int keys because of ambiguity of how
the int is represented from the incoming message. A server might
choose to have int8 keys whereas another server might use int16 keys.
Since we know that rbkit uses int64 keys always, we use that knowledge
to define a int64 map key reader and use that in the rbkit-go code.
This also requires some customization in the generated code where we
need to replace all string map key conversions to int map key
conversions and comparisions.